serialized tests
These DB tests need to be executed in a serialized manner because they each try to touch the DB in different threads at the same time, which either corrupts the process of each test or has one test lock the DB, while another is refused access, which causes it to fail. Either way--we need these tests to execute one after the other (serialized).